requirements.txt
google-generativeai
google-genai
openai
anthropic
python-dotenv
requests
jsonschema
streamlit
